+++ src/aux.cpp
@@ -16,7 +16,7 @@ float strToFloat (std::string inputStr_) {
     return value;
 }
 
-bool member (const std::vector<int> & spots, const int & spot) {
+bool member (const std::vector<unsigned short int> & spots, const int & spot) {
     for (auto i(0u); i < spots.size(); ++i) {
         if (spots[i] == spot)
             return true;
